+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentAndroidEventTest.kt
@@ -24,7 +24,7 @@ class FragmentAndroidEventTest {
                         configureActivity = {
                             initialContract = TestLifecycleContract()
                         },
-                        contracts = {
+                        contracts =  {
                             bind(TestLifecycleContract::class) { _ ->
                                 activityResults().flatMap {
                                     activityResults.add(it)