+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentLifecycleTest.kt
@@ -28,7 +28,7 @@ class FragmentLifecycleTest {
                         contract = TestLifecycleContract()
                         initialContract = contract
                     },
-                    contracts = {
+                    contracts =  {
                         bind(TestLifecycleContract::class) { _ ->
                             Observable.empty()
                         }