+++ formula-android/src/main/java/com/instacart/formula/integration/FlowEnvironment.kt
@@ -1,0 +1,5 @@
+package com.instacart.formula.integration
+
+data class FlowEnvironment<Key>(
+    val onScreenError: (Key, Throwable) -> Unit = { _, it -> throw it }
+)