+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Updated AndroidX fragment-ktx extensions to 1.2.1
 - Bugfix: Fix `Stream.onTerminate` causing illegal state exception.
 - **Breaking**: In formula-android, replacing lastEntry() with visibleState() in FragmentFlowState.
+- Allow down-casting `ActivityStoreContext<MyActivity>` to `ActivityStoreContext<FragmentActivity>`
+- Enable global fragment error logging.
 
 ## [0.5.3] - December 10, 2019
 - Change child formula key from String to Any.